Added MATH 2270 : Final Project
Added the PDF that contains all the details about the final project for this course. I am working on Option A, which requires detailed analysis of PageRank, Google's Search Algorithm, via a Cornell Lecture. I plan to add the presentation and MP4 (if it allows it without a GIT LFS issue again) in the future once I have finished this project.